confirm before deleting post

+++ admin/functions.php
@@ -192,7 +192,7 @@ function find_all_posts() {
             echo "<td>$post_date</td>";   
             echo "<td style='text-align: center'><a href='../post.php?p_id={$post_id}'><i class='fa fa-eye' aria-hidden='true' data-toggle='tooltip' title='View Post'></i></a></td>";  
             echo "<td style='text-align: center'><a href='posts.php?source=edit_post&p_id=$post_id'><i class='fa fa-pencil-square-o' aria-hidden='true' data-toggle='tooltip' title='Edit Post'></i></a></td>";                
-            echo "<td style='text-align: center'><a href='posts.php?delete=$post_id'><i class='fa fa-trash' aria-hidden='true' data-toggle='tooltip' title='Delete Post'></i></a></td>";
+            echo "<td onClick=\"javascript: return confirm('Are you sure you want to delete this post?');\" style='text-align: center'><a href='posts.php?delete=$post_id'><i class='fa fa-trash trash-post' aria-hidden='true' data-toggle='tooltip' title='Delete Post'></i></a></td>";
 
         echo "</tr>";
     }
